backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-multipart

faiss-cpu
pypdf

sentence-transformers==2.2.2
huggingface_hub==0.14.1

langchain==0.1.14
langchain-community==0.0.30

